requirements.txt
pandas
python-docx
PyPDF2
Pillow
scikit-learn
nltk
spaCy
tensorflow
tdqm